add interface for ExchangeServiceClient
